backend/config/requirements.txt
mysqlclient==2.1.1
pika==1.3.1
djangorestframework==3.14.0
django-mysql==4.8.0
django-cors-headers==3.13.0
flask
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.20
Flask-Migrate==2.5.3
Flask-Script==2.0.6
requests==2.28.1
Flask-Cors==3.0.9
markupsafe==2.0.1
itsdangerous<2.0.1
jinja2<3.1.0
werkzeug<2.0.1
pytz==2022.7.1
python_dotenv

